Upgraded bindgen version

+++ ibverbs/src/lib.rs
@@ -1152,7 +1152,7 @@ impl ProtectionDomain {
             ffi::ibv_reg_mr(
                 self.pd,
                 data.as_mut_ptr() as *mut _,
-                n * mem::size_of::<T>(),
+                (n * mem::size_of::<T>()) as u64,
                 access.0 as i32,
             )
         };
